Check EOS Txs for a wallet
A web page that uses the etherscan API to see the transaction data of a specific address for a specific token (EOS). 

Currently only outputs the first Tx as I need to research how to programmatically output all of the Txs to the page without having to hand-code for each one.